pyproject.toml
[build-system]
requires = ["flit_core >=3.3,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "ibis_heavyai"
version = "1.0"
requires-python = ">=3.7"
authors = [{name = "Heavy.AI", email = "community@heavy.ai"}]
description = "Ibis HeavyDB backend"
readme = "README.md"
license = {file = "LICENSE.md"}
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Database",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "ibis-framework >=2.0",
    "heavyai",
    "rbc-project >=0.8.0",
    "pyarrow >=3.0.0",
    "regex",
    "pandas",
    "shapely",
    "geopandas",
    "importlib_metadata",  # drop for python >3.7
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "mock",
]

dev = [
    "heavyai[test]",
    "pre-commit",
    "flit"
]

[project.urls]
Home = "https://github.com/heavyai/ibis-heavyai"
Source = "https://github.com/heavyai/ibis-heavyai"

[tool.flit.sdist]
exclude = [
    "tests/*", "ci/*", ".github/*",
    "*.yaml", ".*"
]

[tool.black]
line-length = 79
skip-string-normalization = true
target-version = ["py37", "py38", "py39", "py310"]
exclude = "docs"

[tool.isort]
ensure_newline_before_comments = true
line_length = 79
multi_line_output = 3
include_trailing_comma = true

[tool.mypy]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rsx -v"
testpaths = [
    "tests",
]
markers = [
    "backend",
    "udf",
    "geo",
    "xfail_backends",
    "xfail_unsupported",
    "skip_backends",
    "xpass_backends",
    "only_on_backends",
    "skip_missing_feature",
    "omniscidb",
    "min_spark_version",
]
